chore: rename sharedRoutes to commonRoutes and modalRoutes
