+++ src/app/Navigation/AuthenticatedRoutes/InboxTab.tsx
@@ -1,5 +1,5 @@
 import { modules } from "app/AppRegistry"
-import { registerSharedRoutes } from "app/Navigation/AuthenticatedRoutes/SharedRoutes"
+import { commonRoutes } from "app/Navigation/AuthenticatedRoutes/CommonRoutes"
 import { registerScreen, StackNavigator } from "app/Navigation/AuthenticatedRoutes/StackNavigator"
 
 export const InboxTab: React.FC = () => {
@@ -14,7 +14,7 @@ export const InboxTab: React.FC = () => {
         module: modules["Conversation"],
       })}
 
-      {registerSharedRoutes()}
+      {commonRoutes()}
     </StackNavigator.Navigator>
   )
 }
