+++ src/app/Navigation/AuthenticatedRoutes/commonRoutes.tsx
@@ -1,6 +1,5 @@
-import { AppModule, modules as allModules } from "app/AppRegistry"
+import { modules as allModules, AppModule } from "app/AppRegistry"
 import { registerScreen, StackNavigator } from "app/Navigation/AuthenticatedRoutes/StackNavigator"
-import { AuthenticatedRoutesStack } from "app/Navigation/AuthenticatedRoutes/Tabs"
 import { isModalScreen } from "app/Navigation/Utils/isModalScreen"
 
 const modules = Object.fromEntries(
@@ -16,10 +15,12 @@ const modules = Object.fromEntries(
   })
 )
 
-const modalModules = Object.entries(modules).filter(([_, module]) => isModalScreen(module))
 const nonModalModules = Object.entries(modules).filter(([_, module]) => !isModalScreen(module))
 
-export const registerSharedRoutes = () => {
+/**
+ * This represents the screens that can be accessed from any tab
+ */
+export const commonRoutes = () => {
   return (
     <StackNavigator.Group>
       {nonModalModules.map(([moduleName, module]) => {
@@ -29,18 +30,5 @@ export const registerSharedRoutes = () => {
         })
       })}
     </StackNavigator.Group>
-  )
-}
-
-export const registerSharedModalRoutes = () => {
-  return (
-    <AuthenticatedRoutesStack.Group screenOptions={{ presentation: "modal" }}>
-      {modalModules.map(([moduleName, module]) => {
-        return registerScreen({
-          name: moduleName as AppModule,
-          module: module,
-        })
-      })}
-    </AuthenticatedRoutesStack.Group>
   )
 }
